# api/docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.SocialCreate:
    properties:
      education_name:
        type: string
      education_url:
        type: string
      instagram_name:
        type: string
      instagram_url:
        type: string
      linkedin_name:
        type: string
      linkedin_url:
        type: string
      location_name:
        type: string
      location_url:
        type: string
      telegram_name:
        type: string
      telegram_url:
        type: string
      twitter_name:
        type: string
      twitter_url:
        type: string
      user_id:
        type: string
      website_name:
        type: string
      website_url:
        type: string
      youtube_name:
        type: string
      youtube_url:
        type: string
    type: object
  models.SocialResponse:
    properties:
      education_name:
        type: string
      education_url:
        type: string
      instagram_name:
        type: string
      instagram_url:
        type: string
      linkedin_name:
        type: string
      linkedin_url:
        type: string
      location_name:
        type: string
      location_url:
        type: string
      telegram_name:
        type: string
      telegram_url:
        type: string
      twitter_name:
        type: string
      twitter_url:
        type: string
      user_id:
        type: string
      website_name:
        type: string
      website_url:
        type: string
      youtube_name:
        type: string
      youtube_url:
        type: string
    type: object
  models.SocialUpdate:
    properties:
      education_name:
        type: string
      education_url:
        type: string
      instagram_name:
        type: string
      instagram_url:
        type: string
      linkedin_name:
        type: string
      linkedin_url:
        type: string
      location_name:
        type: string
      location_url:
        type: string
      telegram_name:
        type: string
      telegram_url:
        type: string
      twitter_name:
        type: string
      twitter_url:
        type: string
      user_id:
        type: string
      website_name:
        type: string
      website_url:
        type: string
      youtube_name:
        type: string
      youtube_url:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      avatar:
        type: string
      bio:
        type: string
      birth_day:
        type: string
      coint:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      refresh_token:
        type: string
      score:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.UserUpdate:
    properties:
      avatar:
        type: string
      bio:
        type: string
      birth_day:
        type: string
      id:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  title: Admin-Monolight
paths:
...
  /v1/social:
    post:
      consumes:
      - application/json
      description: This Api for creating a new social
      parameters:
      - description: SocialCreate Model
        in: body
        name: SocialCreate
        required: true
        schema:
          $ref: '#/definitions/models.SocialCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SocialResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Create Social
      tags:
      - social
    put:
      consumes:
      - application/json
      description: This Api for updating social
      parameters:
      - description: Update Social Model
        in: body
        name: Social
        required: true
        schema:
          $ref: '#/definitions/models.SocialUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SocialResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update Social
      tags:
      - social
  /v1/social/{id}:
    get:
      consumes:
      - application/json
      description: This Api for get social
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SocialResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get Social
      tags:
      - social
...
  /v1/user:
    put:
      consumes:
      - application/json
      description: This Api for updating user
      parameters:
      - description: Update User Model
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - users
  /v1/user/{id}:
    delete:
      consumes:
      - application/json
      description: This Api for deleting user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - users
    get:
      consumes:
      - application/json
      description: This Api for get user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get User
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"